# Add `docstring_limit` parameter to `lambdify` to improve performance for large expressions