--- a/sympy/utilities/lambdify.py
+++ b/sympy/utilities/lambdify.py
@@ -180,7 +180,7 @@ _lambdify_generated_counter = 1
 
 @doctest_depends_on(modules=('numpy', 'scipy', 'tensorflow',), python_version=(3,))
 def lambdify(args, expr, modules=None, printer=None, use_imps=True,
-             dummify=False, cse=False):
+             dummify=False, cse=False, docstring_limit=100):
     """Convert a SymPy expression into a function that allows for fast
     numeric evaluation.