--- a/sympy/utilities/tests/test_lambdify.py
+++ b/sympy/utilities/tests/test_lambdify.py
@@ -1787,6 +1787,7 @@ def test_lambdify_docstring_size_limit_simple_symbol():
         lambdified_expr = lambdify(
             [x],
             x,
+            'sympy',
             docstring_limit=test_case.docstring_limit,
         )
         assert lambdified_expr.__doc__ == test_case.expected_docstring
@@ -1820,6 +1821,7 @@ def test_lambdify_docstring_size_limit_nested_expr():
         lambdified_expr = lambdify(
             [x, y, z],
             expr,
+            'sympy',
             docstring_limit=test_case.docstring_limit,
         )
         assert lambdified_expr.__doc__ == test_case.expected_docstring
@@ -1835,9 +1837,9 @@ def test_lambdify_docstring_size_limit_matrix():
         )
         SRC = (
             'def _lambdifygenerated(x, y, z):\n'
-            '    return array([[0, x], [x + y + z, x**3 + 3*x**2*y + '
-            '3*x**2*z + 3*x*y**2 + 6*x*y*z + 3*x*z**2 + y**3 + 3*y**2*z + '
-            '3*y*z**2 + z**3]])\n'
+            '    return ImmutableDenseMatrix([[0, x], [x + y + z, x**3 + '
+            '3*x**2*y + 3*x**2*z + 3*x*y**2 + 6*x*y*z + 3*x*z**2 + y**3 + '
+            '3*y**2*z + 3*y*z**2 + z**3]])\n'
         )
 
     x, y, z = symbols('x, y, z')
@@ -1854,6 +1856,7 @@ def test_lambdify_docstring_size_limit_matrix():
         lambdified_expr = lambdify(
             [x, y, z],
             expr,
+            'sympy',
             docstring_limit=test_case.docstring_limit,
         )
         assert lambdified_expr.__doc__ == test_case.expected_docstring