<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->
Fixes #24743

#### Brief description of what is fixed or changed
Issue #24743 showed that for large expression, a prohibitively large time is spent producing the string representation of the expression for the doctoring in the lambdified function. This PR adds a `docstring_limit` argument to `lambdify` which allows users to control whether or not the string representation is included in the docstring. 

Behaviour:
- If `docstring_limit=None` then `lambdify` prints the docstring with full expression and source code.
- If `docstring_limit` is a strictly positive `int` and the the `docstring_limit` is not exceeded then the docstring is printed with full expression and source code.
- If `docstring_limit` is `0`, or is a negative `int`, or is a strictly positive `int` and the number of nodes in the expression exceeds `docstring_limit` then `lambdify` includes messages in the docstring that the expression and source code have been redacted due to length.
The default is `docstring_limit=1000`.

#### Other comments
For [this test case on the original issue](https://github.com/sympy/sympy/issues/24743#issuecomment-1436597104), an exponential performance improvement can be seen as expressions grow larger. For different values of `N` simple profile results are (LHS is `docstring_limit=None`, RHS is `docstring_limit=0`):
```
N = 1: 0.191129 s -> 0.18506 s
N = 2: 0.216199 s -> 0.20614 s
N = 3: 0.379411 s -> 0.289574 s
N = 4: 3.72576 s -> 0.800366 s
N = 5: 116.388 s -> 15.1056 s
```

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave `-log(x)`.

* physics.units
  * Corrected a semantical error in the conversion between volt and statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* utilities
  * Added a new parameter, `docstring_limit`, to `lambdify` that allows users to control the size of expressions that are included as fully rendered in the docstring.
<!-- END RELEASE NOTES -->