@brocksam 2023-02-21
Test failure on `sympy/utilities/tests/test_lambdify.py:test_lambdify_docstring_size_limit_matrix` (new test with this PR). Interestingly, I can run the test suite locally using `pytest`, `bin/test`, and `bin/test -s` and all pass so not sure what's causing the failure here.