@brocksam 2023-02-21
> Test failure on `sympy/utilities/tests/test_lambdify.py:test_lambdify_docstring_size_limit_matrix` (new test with this PR). Interestingly, I can run the test suite locally using `pytest`, `bin/test`, and `bin/test -s` and all pass so not sure what's causing the failure here.

Resolve in https://github.com/sympy/sympy/pull/24754/commits/e5c155550e7bfd38133d8554bb58961b4243486e. Testing locally I had NumPy installed, whereas in CI NumPy is not installed. `lambdify` will use `numpy` as the backend if NumPy is available, but will default to the `sympy` backend in NumPy cannot be imported. This results in `Matrix` being printed as `ImmutableDenseMatrix(...)` with the `sympy` backend versus `array(...)` being printed with the `numpy` backend. The solution was to call `lambdify` in the tests with `sympy` passed as an argument to the `modules` parameter.