@brocksam 2023-02-21
> I just tested this out and it works nicely. I like the default 1000 limit. It should let most things through that a user would desire to read. Great speed up! Could be worth adding a benchmark as a regression indicator in the benchmarks repo.

I agree, I will add the test case by @TJStienstra from the original issue, which I present profile results from above, to the benchmarks repo.