@TJStienstra 2023-03-07
> Would it have been enough to just use `sstr(expr, order='none')` here instead? lambdify is already printing the expression (using the code printers), the str printer really shouldn't be any slower.

Just did the experiment from [this comment](https://github.com/sympy/sympy/issues/24743#issuecomment-1436597104) with a profiler to compare speeds. This showed that `sstr(expr, order='none')` is indeed faster than `str`, but it still significantly slows down `lambdify` in the order of 35%.