@moorepants 2023-03-07
@asmeurer If you want it changed, then open a PR before 1.12 gets released. Sorry it went in before you could review.