@asmeurer 2023-03-07
The change here is fine to keep if it's useful, but I don't think it's a good idea to pollute the API with a new option that doesn't really provide any benefit, and would need to be removed if we ever fixed the underlying performance issue.