@bjodah 2023-03-07
We can make it non-public? And maybe up the default from 1000 to something bigger (10000?) as long as the redaction message contains the information needed for how to update (increase) the limit.