@asmeurer 2023-03-07
Yes, my suggestion was to configure it with a global variable in lambdify.py. That way users can still modify it if they really want to, but it also isn't blatantly in users' faces as if it was that important. We could even start it with an underscore to make it clear that it could go away.