@moorepants 2023-03-08
> would need to be removed if we ever fixed the underlying performance issue

I'm not seeing how it is just a performance issue. We are lambdifying expressions with millions of operations and, for those, we should spend zero compute time on printing anything in the docstring even if you can improve the performance.